research_agent/requirements.txt
"""
Requirements for Unified Research Agent System
"""

# Core LLM & Agent Framework
smolagents==1.23.0
litellm>=1.0.0

# API Clients
requests>=2.31.0
python-dotenv>=1.0.0

# Web & Browser
selenium>=4.0.0
beautifulsoup4>=4.12.0

# Document Generation
python-docx>=0.8.11

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Research & Search
tavily-python>=0.3.0

# Utilities
pydantic>=2.0.0
dspy-ai>=2.4.0

# Optional: For async support
aiohttp>=3.8.0
asyncio-contextmanager>=1.0.0
